feat(appointments): add golden tests and refactor widgets
- Extract `AppointmentCard` and `AppointmentForm` to standalone widgets.
- Add golden tests for `AppointmentsPage`, `AppointmentCard`, and `AppointmentForm`.
- Clean up stale legacy tests in `test/features/appointments/data/`.
- Fix repository method calls in `GetAllAppointmentsUseCase` tests.
