+++ lib/features/appointments/presentation/pages/appointments_page.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:isar/isar.dart';
 import 'package:intl/intl.dart';
 import '../../../../core/di/injection.dart';
 import '../../../../core/theme/cyber_theme.dart';
@@ -8,6 +7,8 @@ import '../../../email-citas/presentation/email_connect_page.dart';
 import '../../../calendar_import/presentation/calendar_import_page.dart';
 import '../../domain/entities/appointment.dart';
 import '../../domain/repositories/appointment_repository.dart';
+import '../widgets/appointment_card.dart';
+import '../widgets/appointment_form.dart';
 
 class AppointmentsPage extends StatefulWidget {
   const AppointmentsPage({super.key});
@@ -124,7 +125,7 @@ class _AppointmentsPageState extends State<AppointmentsPage> {
                         )
                       : SliverList(
                           delegate: SliverChildBuilderDelegate(
-                            (context, index) => _AppointmentCard(
+                            (context, index) => AppointmentCard(
                               appointment: upcoming[index],
                               onTap: () => _showAppointmentForm(appointment: upcoming[index]),
                             ),
@@ -149,7 +150,7 @@ class _AppointmentsPageState extends State<AppointmentsPage> {
                         )
                       : SliverList(
                           delegate: SliverChildBuilderDelegate(
-                            (context, index) => _AppointmentCard(
+                            (context, index) => AppointmentCard(
                               appointment: past[index],
                               onTap: () => _showAppointmentForm(appointment: past[index]),
                             ),
@@ -265,7 +266,7 @@ class _AppointmentsPageState extends State<AppointmentsPage> {
       context: context,
       isScrollControlled: true,
       backgroundColor: Colors.transparent,
-      builder: (context) => _AppointmentForm(
+      builder: (context) => AppointmentForm(
         appointment: appointment,
         onSave: (newApp) async {
           await _repository.saveAppointment(newApp);
@@ -275,261 +276,6 @@ class _AppointmentsPageState extends State<AppointmentsPage> {
           await _repository.deleteAppointment(id);
           _loadAppointments();
         },
-      ),
-    );
-  }
-}
-
-class _AppointmentCard extends StatelessWidget {
-  final Appointment appointment;
-  final VoidCallback onTap;
-
-  const _AppointmentCard({required this.appointment, required this.onTap});
-
-  @override
-  Widget build(BuildContext context) {
-    final color = appointment.status == AppointmentStatus.upcoming
-        ? Colors.teal
-        : (appointment.status == AppointmentStatus.completed ? Colors.green : Colors.grey);
-
-    return Padding(
-      padding: const EdgeInsets.symmetric(horizontal: 16.0, vertical: 6.0),
-      child: GlassmorphicCard(
-        child: InkWell(
-          onTap: onTap,
-          borderRadius: BorderRadius.circular(16),
-          child: Padding(
-            padding: const EdgeInsets.all(16.0),
-            child: Row(
-              children: [
-                Expanded(
-                  child: Column(
-                    mainAxisSize: MainAxisSize.min,
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    children: [
-                      Text(
-                        appointment.doctorName,
-                        style: const TextStyle(fontWeight: FontWeight.bold, fontSize: 16),
-                        overflow: TextOverflow.ellipsis,
-                      ),
-                      const SizedBox(height: 2),
-                      Text(
-                        appointment.specialty,
-                        style: const TextStyle(color: Colors.white70, fontSize: 14),
-                        overflow: TextOverflow.ellipsis,
-                      ),
-                      const SizedBox(height: 8),
-                      Wrap(
-                        spacing: 12,
-                        runSpacing: 4,
-                        crossAxisAlignment: WrapCrossAlignment.center,
-                        children: [
-                          Row(
-                            mainAxisSize: MainAxisSize.min,
-                            children: [
-                              const Icon(Icons.access_time, size: 14, color: CyberTheme.secondary),
-                              const SizedBox(width: 4),
-                              Flexible(
-                                child: Text(
-                                  DateFormat('dd MMM, hh:mm a', 'es').format(appointment.dateTime),
-                                  style: const TextStyle(color: CyberTheme.secondary, fontSize: 12),
-                                  overflow: TextOverflow.ellipsis,
-                                ),
-                              ),
-                            ],
-                          ),
-                          if (appointment.source != null)
-                            Row(
-                              mainAxisSize: MainAxisSize.min,
-                              children: [
-                                const Icon(Icons.email_outlined, size: 14, color: Colors.white54),
-                                const SizedBox(width: 4),
-                                Flexible(
-                                  child: Text(
-                                    appointment.source!,
-                                    style: const TextStyle(color: Colors.white54, fontSize: 12),
-                                    overflow: TextOverflow.ellipsis,
-                                  ),
-                                ),
-                              ],
-                            ),
-                        ],
-                      ),
-                    ],
-                  ),
-                ),
-                const SizedBox(width: 8),
-                Container(
-                  padding: const EdgeInsets.symmetric(horizontal: 10, vertical: 4),
-                  decoration: BoxDecoration(
-                    color: color.withValues(alpha: 0.1),
-                    borderRadius: BorderRadius.circular(8),
-                    border: Border.all(color: color.withValues(alpha: 0.5)),
-                  ),
-                  child: Text(
-                    appointment.status.name.toUpperCase(),
-                    style: TextStyle(color: color, fontSize: 10, fontWeight: FontWeight.bold),
-                  ),
-                ),
-              ],
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class _AppointmentForm extends StatefulWidget {
-  final Appointment? appointment;
-  final Function(Appointment) onSave;
-  final Function(int) onDelete;
-
-  const _AppointmentForm({this.appointment, required this.onSave, required this.onDelete});
-
-  @override
-  State<_AppointmentForm> createState() => _AppointmentFormState();
-}
-
-class _AppointmentFormState extends State<_AppointmentForm> {
-  late TextEditingController _doctorController;
-  late TextEditingController _specialtyController;
-  late TextEditingController _notesController;
-  late DateTime _selectedDate;
-  late TimeOfDay _selectedTime;
-  late AppointmentStatus _status;
-
-  @override
-  void initState() {
-    super.initState();
-    _doctorController = TextEditingController(text: widget.appointment?.doctorName ?? '');
-    _specialtyController = TextEditingController(text: widget.appointment?.specialty ?? '');
-    _notesController = TextEditingController(text: widget.appointment?.notes ?? '');
-    _selectedDate = widget.appointment?.dateTime ?? DateTime.now();
-    _selectedTime = TimeOfDay.fromDateTime(_selectedDate);
-    _status = widget.appointment?.status ?? AppointmentStatus.upcoming;
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      decoration: const BoxDecoration(
-        color: CyberTheme.surfaceDark,
-        borderRadius: BorderRadius.vertical(top: Radius.circular(20)),
-      ),
-      padding: EdgeInsets.only(
-        bottom: MediaQuery.of(context).viewInsets.bottom,
-        left: 20,
-        right: 20,
-        top: 20,
-      ),
-      child: SingleChildScrollView(
-        child: Column(
-          mainAxisSize: MainAxisSize.min,
-          crossAxisAlignment: CrossAxisAlignment.stretch,
-          children: [
-            Text(
-              widget.appointment == null ? 'Nueva Cita' : 'Editar Cita',
-              style: const TextStyle(fontSize: 20, fontWeight: FontWeight.bold),
-              textAlign: TextAlign.center,
-            ),
-            const SizedBox(height: 20),
-            TextField(
-              controller: _doctorController,
-              decoration: const InputDecoration(labelText: 'Nombre del Doctor', border: OutlineInputBorder()),
-            ),
-            const SizedBox(height: 12),
-            TextField(
-              controller: _specialtyController,
-              decoration: const InputDecoration(labelText: 'Especialidad', border: OutlineInputBorder()),
-            ),
-            const SizedBox(height: 12),
-            Row(
-              children: [
-                Expanded(
-                  child: Material(
-                    color: Colors.transparent,
-                    child: ListTile(
-                      title: const Text('Fecha'),
-                      subtitle: Text(DateFormat('yyyy-MM-dd').format(_selectedDate)),
-                      onTap: () async {
-                        final picked = await showDatePicker(
-                          context: context,
-                          initialDate: _selectedDate,
-                          firstDate: DateTime(2000),
-                          lastDate: DateTime(2100),
-                        );
-                        if (picked != null) setState(() => _selectedDate = picked);
-                      },
-                    ),
-                  ),
-                ),
-                Expanded(
-                  child: Material(
-                    color: Colors.transparent,
-                    child: ListTile(
-                      title: const Text('Hora'),
-                      subtitle: Text(_selectedTime.format(context)),
-                      onTap: () async {
-                        final picked = await showTimePicker(
-                          context: context,
-                          initialTime: _selectedTime,
-                        );
-                        if (picked != null) setState(() => _selectedTime = picked);
-                      },
-                    ),
-                  ),
-                ),
-              ],
-            ),
-            const SizedBox(height: 12),
-            DropdownButtonFormField<AppointmentStatus>(
-              initialValue: _status,
-              decoration: const InputDecoration(labelText: 'Estado', border: OutlineInputBorder()),
-              items: AppointmentStatus.values.map((s) => DropdownMenuItem(value: s, child: Text(s.name))).toList(),
-              onChanged: (val) => setState(() => _status = val!),
-            ),
-            const SizedBox(height: 12),
-            TextField(
-              controller: _notesController,
-              maxLines: 3,
-              decoration: const InputDecoration(labelText: 'Notas (Opcional)', border: OutlineInputBorder()),
-            ),
-            const SizedBox(height: 24),
-            ElevatedButton(
-              onPressed: () {
-                final finalDateTime = DateTime(
-                  _selectedDate.year,
-                  _selectedDate.month,
-                  _selectedDate.day,
-                  _selectedTime.hour,
-                  _selectedTime.minute,
-                );
-                final app = Appointment(
-                  id: widget.appointment?.id ?? Isar.autoIncrement,
-                  doctorName: _doctorController.text,
-                  specialty: _specialtyController.text,
-                  dateTime: finalDateTime,
-                  notes: _notesController.text,
-                  status: _status,
-                );
-                widget.onSave(app);
-                Navigator.pop(context);
-              },
-              style: ElevatedButton.styleFrom(backgroundColor: CyberTheme.primary, foregroundColor: Colors.black),
-              child: const Text('GUARDAR'),
-            ),
-            if (widget.appointment != null)
-              TextButton(
-                onPressed: () {
-                  widget.onDelete(widget.appointment!.id);
-                  Navigator.pop(context);
-                },
-                child: const Text('ELIMINAR', style: TextStyle(color: Colors.redAccent)),
-              ),
-            const SizedBox(height: 20),
-          ],
-        ),
       ),
     );
   }
